requirements.txt
Django==2.2.1
django-tinymce==3.5.0
Pillow==8.2.0
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.4.0
